requirements.txt
fastapi
firebase_admin
uvicorn[standard]
gunicorn
numpy
pandas
python-socketio
aiohttp
aiodns